🔨 Refactor(#127): useUserRankingQuery(+API) user 도메인으로 구조 변경
